requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-mathfilters==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
h11==0.13.0
idna==3.3
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-decouple==3.6
pytz==2022.6
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
